added:
  daemon/system/System/ignore.hpp
  daemon/system/System/ignore.t.cpp
  doc/devel/creating_filters.txt
  doc/devel/creating_inputs.txt
  doc/devel/creating_triggers.txt
  doc/example_configs/
  doc/example_configs/example1.xml
  doc/example_configs/minimal.xml
  wui/site/OpenLayers.js
  wui/site/pics/OSL/
  wui/site/pics/OSL/add_point_off.png
  wui/site/pics/OSL/add_point_on.png
  wui/site/pics/OSL/blank.gif
  wui/site/pics/OSL/close.gif
  wui/site/pics/OSL/drag-rectangle-off.png
  wui/site/pics/OSL/drag-rectangle-on.png
  wui/site/pics/OSL/draw_line_off.png
  wui/site/pics/OSL/draw_line_on.png
  wui/site/pics/OSL/draw_point_off.png
  wui/site/pics/OSL/draw_point_on.png
  wui/site/pics/OSL/draw_polygon_off.png
  wui/site/pics/OSL/draw_polygon_on.png
  wui/site/pics/OSL/editing_tool_bar.png
  wui/site/pics/OSL/move_feature_off.png
  wui/site/pics/OSL/move_feature_on.png
  wui/site/pics/OSL/navigation_history.png
  wui/site/pics/OSL/overview_replacement.gif
  wui/site/pics/OSL/pan-panel-NOALPHA.png
  wui/site/pics/OSL/pan-panel.png
  wui/site/pics/OSL/pan_off.png
  wui/site/pics/OSL/pan_on.png
  wui/site/pics/OSL/panning-hand-off.png
  wui/site/pics/OSL/panning-hand-on.png
  wui/site/pics/OSL/remove_point_off.png
  wui/site/pics/OSL/remove_point_on.png
  wui/site/pics/OSL/ruler.png
  wui/site/pics/OSL/save_features_off.png
  wui/site/pics/OSL/save_features_on.png
  wui/site/pics/OSL/view_next_off.png
  wui/site/pics/OSL/view_next_on.png
  wui/site/pics/OSL/view_previous_off.png
  wui/site/pics/OSL/view_previous_on.png
  wui/site/pics/OSL/zoom-panel-NOALPHA.png
  wui/site/pics/OSL/zoom-panel.png
  wui/site/pics/destinations.png
  wui/site/pics/sources.png
  wui/site/protected/common/Pair.php
  wui/site/protected/common/Sortedheatmap.php
  wui/site/styleOL.css
modified:
  daemon/commons/Commons/Filesystem/openElement.cpp
  daemon/commons/Commons/Filesystem/openElement.t.cpp
  daemon/commons/Commons/SharedPtrNotNULL.hpp
  daemon/core/Core/Processors.cpp
  daemon/core/Core/Sources.cpp
  daemon/filter/Filter/Simple/Strategy.hpp
  daemon/input/Input/Thread.cpp
  daemon/inputfile/Input/File/IStreamReader.cpp
  daemon/inputfile/Input/File/IStreamReader.t.cpp
  daemon/inputfile/Input/File/ParserCallback.cpp
  daemon/inputfile/Input/File/ParserCallback.t.cpp
  daemon/rfcio/RFCIO/FromXML.cpp
  daemon/system/System/ScopedPtrCustom.hpp
  daemon/system/System/SharedPtrNotNULL.hpp
  daemon/system/System/SharedPtrNotNULL.t.cpp
  daemon/trigger/Trigger/Compose/Full.t.cpp
  daemon/trigger/Trigger/Compose/LinkWUI.cpp
  daemon/trigger/Trigger/Compose/LinkWUI.t.cpp
  daemon/trigger/Trigger/Compose/Summary.t.cpp
  data_model/postgres/8.3/06_drop_meta_alerts.sql
  data_model/postgres/8.3/11_create_wui.sql
  data_model/postgres/8.3/12_create_alerts.sql
  data_model/postgres/8.3/13_create_meta_alerts.sql
  data_model/postgres/8.3/14_create_processor_states.sql
  data_model/postgres/8.4/06_drop_meta_alerts.sql
  data_model/postgres/8.4/11_create_wui.sql
  data_model/postgres/8.4/12_create_alerts.sql
  data_model/postgres/8.4/13_create_meta_alerts.sql
  data_model/postgres/8.4/14_create_processor_states.sql
  doc/INSTALLATION.txt
  wui/site/protected/application.xml
  wui/site/protected/common/ImageService.php
  wui/site/protected/common/Layout.tpl
  wui/site/protected/common/RangeSelector.tpl
  wui/site/protected/mp-sqlmap-postgres.xml
  wui/site/protected/pages/DataMiningHeatMap.page
  wui/site/protected/pages/DataMiningHeatMap.php
  wui/site/protected/pages/DataMiningSeverity.php
  wui/site/protected/pages/MetaAlert.php
  wui/site/style.css
pending merges:
  Bartek 'BaSz' Szu... 2010-11-17 added info on WUI installation to INSTALLATION.txt manual.
    Bartek 'BaSz' Szu... 2010-11-17 [merge] rebase.
    Wojciech Waga 2010-11-17 [merge] rebase
    Wojciech Waga 2010-11-17 ts
    Wojciech Waga 2010-11-17 ts
    Wojciech Waga 2010-11-17 [merge] rebase
    Wojciech Waga 2010-11-17 WUI/HeatMap
    Bartek 'BaSz' Szu... 2010-11-17 timeout for reading from input has been enlarged to 45[s].
    Bartek 'BaSz' Szu... 2010-11-17 implemented interruptable timeout for input::file.
    Bartek 'BaSz' Szu... 2010-11-17 fixed link to use valid aprameter.
    Bartek 'BaSz' Szu... 2010-11-17 commente aded.
    Bartek 'BaSz' Szu... 2010-11-17 updated version of sharedptrnn added.
    Bartek 'BaSz' Szu... 2010-11-17 changed some important logs from DEBUG to INFO so that they'll be present in...
    Bartek 'BaSz' Szu... 2010-11-16 updated DB schemas for PSQL to use 64-bit integers as indexes.
    Bartek 'BaSz' Szu... 2010-11-16 example config files added.
    Bartek 'BaSz' Szu... 2010-11-16 info on changing ID is now WARN not ERROR.
    Bartek 'BaSz' Szu... 2010-11-16 fixed issue with reading files that have no write-access.
    Bartek 'BaSz' Szu... 2010-11-16 added missing test on openElement call;
    Bartek 'BaSz' Szu... 2010-11-16 typo fixed.
    Bartek 'BaSz' Szu... 2010-11-16 added more TODOs on documentation.
    Bartek 'BaSz' Szu... 2010-11-15 fixed tests for changed URLs.
    Bartek 'BaSz' Szu... 2010-11-15 new, development version of System library included.

+++ data_model/postgres/8.3/12_create_alerts.sql
@@ -18,7 +18,7 @@ CREATE TABLE    severities
 CREATE SEQUENCE alerts_id_seq;
 CREATE TABLE    alerts
 (
-  id          int          PRIMARY KEY DEFAULT nextval('alerts_id_seq'),
+  id          bigint       PRIMARY KEY DEFAULT nextval('alerts_id_seq'),
   name        varchar(256) NOT NULL,
   detect_time timestamp    NULL,
   create_time timestamp    NOT NULL DEFAULT now(),
@@ -38,7 +38,7 @@ CREATE TABLE    alerts
 CREATE SEQUENCE reference_urls_id_seq;
 CREATE TABLE    reference_urls
 (
-  id   int          PRIMARY KEY DEFAULT nextval('reference_urls_id_seq'),
+  id   bigint       PRIMARY KEY DEFAULT nextval('reference_urls_id_seq'),
   name varchar(128) NOT NULL,
   url  varchar(256) NOT NULL
 );
@@ -48,9 +48,9 @@ CREATE TABLE    reference_urls
 CREATE SEQUENCE hosts_id_seq;
 CREATE TABLE    hosts
 (
-  id       int      PRIMARY KEY DEFAULT nextval('hosts_id_seq'),
-  id_alert int      NOT NULL REFERENCES alerts(id),
-  id_ref   int      NULL REFERENCES reference_urls(id),
+  id       bigint   PRIMARY KEY DEFAULT nextval('hosts_id_seq'),
+  id_alert bigint   NOT NULL REFERENCES alerts(id),
+  id_ref   bigint   NULL REFERENCES reference_urls(id),
   role     char(3)  NOT NULL,
   ip       inet     NOT NULL,
   mask     inet     NULL,
@@ -65,9 +65,9 @@ CREATE TABLE    hosts
 CREATE SEQUENCE procs_id_seq;
 CREATE TABLE    procs
 (
-  id        int          PRIMARY KEY DEFAULT nextval('procs_id_seq'),
-  id_host   int          NOT NULL REFERENCES hosts(id),
-  id_ref    int          NULL REFERENCES reference_urls(id),
+  id        bigint       PRIMARY KEY DEFAULT nextval('procs_id_seq'),
+  id_host   bigint       NOT NULL REFERENCES hosts(id),
+  id_ref    bigint       NULL REFERENCES reference_urls(id),
   path      varchar(256) NULL,
   name      varchar(256) NOT NULL,
   md5       char(32)     NULL,
@@ -83,9 +83,9 @@ CREATE INDEX procs_id_host_index ON procs(id_host);
 CREATE SEQUENCE services_id_seq;
 CREATE TABLE    services
 (
-  id       int      PRIMARY KEY DEFAULT nextval('services_id_seq'),
-  id_host  int      NOT NULL REFERENCES hosts(id),
-  id_ref   int      NULL REFERENCES reference_urls(id),
+  id       bigint   PRIMARY KEY DEFAULT nextval('services_id_seq'),
+  id_host  bigint   NOT NULL REFERENCES hosts(id),
+  id_ref   bigint   NULL REFERENCES reference_urls(id),
   name     char(32) NOT NULL,
   port     int      NOT NULL,
   protocol char(32) NULL,
@@ -98,8 +98,8 @@ CREATE TABLE    services
 CREATE SEQUENCE analyzers_id_seq;
 CREATE TABLE    analyzers
 (
-  id      int          PRIMARY KEY DEFAULT nextval('analyzers_id_seq'),
-  sys_id  int          NOT NULL UNIQUE,
+  id      bigint       PRIMARY KEY DEFAULT nextval('analyzers_id_seq'),
+  sys_id  bigint       NOT NULL UNIQUE,
   name    varchar(128) NOT NULL,
   version char(16)     NULL,
   os      varchar(128) NULL,
@@ -113,8 +113,8 @@ CREATE TABLE    analyzers
 -- alerts<->analyzers
 CREATE TABLE alert_analyzers
 (
-  id_alert    int NOT NULL REFERENCES alerts(id),
-  id_analyzer int NOT NULL REFERENCES analyzers(id),
+  id_alert    bigint NOT NULL REFERENCES alerts(id),
+  id_analyzer bigint NOT NULL REFERENCES analyzers(id),
 
   UNIQUE(id_alert, id_analyzer)
 );
